# data/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package data

type Items struct {
	ItemID   int32   `json:"item_id"`
	ItemName string  `json:"item_name"`
	Price    float64 `json:"price"`
	Quantity int32   `json:"quantity"`
}

type Products struct {
	ProductID   int32   `json:"product_id"`
	ProductName string  `json:"product_name"`
	Price       float64 `json:"price"`
	Stock       int32   `json:"stock"`
}


type RequestForAddItem struct {
	ProductID int32 `json:"product_id"`
	Quantity  int32 `json:"quantity"`
}

type RequestForUpdate struct {
	ItemID   int32 `json:"item_id"`
	Quantity int32 `json:"quantity"`
}

type RequestForDiscount struct {
	DiscountType string  `json:"discount_type"`  
	Discount     float64 `json:"discount"`      
	ItemID       int32   `json:"item_id"`       
}